+++ inewave/libs/__init__.py
@@ -1,0 +1,5 @@
+# Inclui os membros
+
+from .eolica import Eolica  # noqa
+from .restricoes import Restricoes  # noqa
+from .usinas_hidreletricas import UsinasHidreletricas  # noqa
