--- tests/mocks/arquivos/eolicaconfig.py
@@ -1,16 +1,0 @@
-MockRegistroEolicaConfiguracaoPeriodo = (
-    "EOLICA-CONFIGURACAO-PERIODO ; 1 ; 2021/01; 2030/12; centralizado\n"
-)
-
-MockRegistroPEEConfiguracaoPeriodo = (
-    "PEE-CONFIG-PER  ; 1         ; 2021/01; 2030/12; centralizado\n"
-)
-
-MockEolicaConfig = [
-    "&EOLICA-CONFIGURACAO-PERIODO; CodigoEolica; PeriodoInicial; PeriodoFinal  ; EstadoEolica (centralizado = 1, fixo = 2, nãoexistente = 3)\n",
-    "EOLICA-CONFIGURACAO-PERIODO ; 1 ; 2021/01; 2030/12; centralizado\n",
-    "EOLICA-CONFIGURACAO-PERIODO ; 2 ; 2021/01; 2030/12; centralizado\n",
-    "EOLICA-CONFIGURACAO-PERIODO ; 3 ; 2021/01; 2030/12; centralizado\n",
-    "EOLICA-CONFIGURACAO-PERIODO ; 4 ; 2021/01; 2030/12; centralizado\n",
-    "EOLICA-CONFIGURACAO-PERIODO ; 5 ; 2021/01; 2030/12; centralizado\n",
-]
