--- tests/mocks/arquivos/eolicafte.py
@@ -1,13 +1,0 @@
-MockRegistroFuncaoProducao = "EOLICA-FUNCAO-PRODUCAO-VENTO-POTENCIA-LINEAR-PU-PERIODO ; 1; 2021/01; 2030/12;  -0.14454132; 0.109046370\n"
-
-MockRegistroPEEFTE = "PEE-FPVP-LIN-PU-PER	  ; 1         ; 2021/01; 2030/12; -0.14454132687670500 ; 0.10904637648150100\n"
-
-MockEolicaFTE = [
-    "& EOLICA-FUNCAO-PRODUCAO-VENTO-POTENCIA-LINEAR-PU-PERIODO ; CodEolica;PerIni ;PerFin ; CoeficienteLinear; CoeficienteAngular\n",
-    "\n",
-    "EOLICA-FUNCAO-PRODUCAO-VENTO-POTENCIA-LINEAR-PU-PERIODO ; 1; 2021/01; 2030/12;  -0.14454132; 0.10904637\n",
-    "EOLICA-FUNCAO-PRODUCAO-VENTO-POTENCIA-LINEAR-PU-PERIODO ; 2; 2021/01; 2030/12;  -0.34282825; 0.09599283\n",
-    "EOLICA-FUNCAO-PRODUCAO-VENTO-POTENCIA-LINEAR-PU-PERIODO ; 3; 2021/01; 2030/12;  -0.31163518; 0.14003760\n",
-    "EOLICA-FUNCAO-PRODUCAO-VENTO-POTENCIA-LINEAR-PU-PERIODO ; 4; 2021/01; 2030/12;  -0.08251197; 0.07183436\n",
-    "EOLICA-FUNCAO-PRODUCAO-VENTO-POTENCIA-LINEAR-PU-PERIODO ; 5; 2021/01; 2030/12;  -0.24408456; 0.09996426",
-]
