--- tests/mocks/arquivos/eolicahistorico.py
@@ -1,31 +1,0 @@
-MockRegistroEolicaHistoricoHorizonte = (
-    "EOLICA-HISTORICO-VENTO-HORIZONTE ; 1979/01 ; 2016/01\n"
-)
-
-MockRegistroEolicaHistorico = (
-    "EOLICA-HISTORICO-VENTO; 1; 1979/01; 1979/02;    3.43; 1.0\n"
-)
-
-MockRegistroHistoricoVentoHorizonte = (
-    "VENTO-HIST-HORIZ  ; 1979/01     ; 2016/01\n"
-)
-
-MockRegistroHistoricoVento = "VENTO-HIST  ; 1           ; 1979/01     ; 1979/02   ;      4.05       ; 1.0\n"
-
-MockEolicaHistorico = [
-    "& EOLICA-HISTORICO-VENTO-HORIZONTE ; DataInicial ; DataFinal\n",
-    "EOLICA-HISTORICO-VENTO-HORIZONTE ; 1979/01 ; 2016/01\n",
-    "\n",
-    "\n",
-    "& EOLICA-HISTORICO-VENTO ; CodigoEolica ; DataInicial ; DataFinal ; MagnitudeVelocidadeVento ; DirecaoVento \n",
-    "EOLICA-HISTORICO-VENTO; 1; 1979/01; 1979/02;    3.43; 1.0\n",
-    "EOLICA-HISTORICO-VENTO; 1; 1979/02; 1979/03;    4.28; 1.0\n",
-    "EOLICA-HISTORICO-VENTO; 1; 1979/03; 1979/04;    4.67; 1.0\n",
-    "EOLICA-HISTORICO-VENTO; 1; 1979/04; 1979/05;    4.50; 1.0\n",
-    "EOLICA-HISTORICO-VENTO; 1; 1979/05; 1979/06;    4.77; 1.0\n",
-    "EOLICA-HISTORICO-VENTO; 1; 1979/06; 1979/07;    6.09; 1.0\n",
-    "EOLICA-HISTORICO-VENTO; 1; 1979/07; 1979/08;    5.66; 1.0\n",
-    "EOLICA-HISTORICO-VENTO; 1; 1979/08; 1979/09;    6.21; 1.0\n",
-    "EOLICA-HISTORICO-VENTO; 1; 1979/09; 1979/10;    6.75; 1.0\n",
-    "EOLICA-HISTORICO-VENTO; 1; 1979/10; 1979/11;    5.64; 1.0\n",
-]
