--- tests/mocks/arquivos/eolicaposto.py
@@ -1,19 +1,0 @@
-MockRegistroPostoCadastro = ("POSTO-VENTO-CAD  ; 1           ; NEInterior\n",)
-
-MockRegistroPEEPosto = "PEE-POSTO  ; 1         ; 1\n"
-
-MockEolicaPosto = [
-    "&POSTO-VENTO-CAD ; CodigoPosto ; NomePosto\n",
-    "POSTO-VENTO-CAD  ; 1           ; NEInterior\n",
-    "POSTO-VENTO-CAD  ; 2           ; NELitoral\n",
-    "POSTO-VENTO-CAD  ; 3           ; NEPE\n",
-    "POSTO-VENTO-CAD  ; 4           ; SULInterior\n",
-    "POSTO-VENTO-CAD  ; 5           ; SULLitoral\n",
-    "\n",
-    "&PEE-POSTO ; CodigoPEE ; CodigoPosto\n",
-    "PEE-POSTO  ; 1         ; 1\n",
-    "PEE-POSTO  ; 2         ; 2\n",
-    "PEE-POSTO  ; 3         ; 3\n",
-    "PEE-POSTO  ; 4         ; 4\n",
-    "PEE-POSTO  ; 5         ; 5\n",
-]
