--- tests/mocks/arquivos/eolicasubmercado.py
@@ -1,13 +1,0 @@
-MockRegistroEolicaSubmercado = "EOLICA-SUBMERCADO ; 5 ; 2\n"
-
-MockRegistroPEESubmercado = "PEE-SUBM   ; 1         ; 3\n"
-
-MockEolicaSubmercado = [
-    "& EOLICA-SUBMERCADO; CodigoEolica; CodigoSubmercado\n",
-    "EOLICA-SUBMERCADO ; 1 ; 3\n",
-    "EOLICA-SUBMERCADO ; 2 ; 3\n",
-    "EOLICA-SUBMERCADO ; 3 ; 3\n",
-    "EOLICA-SUBMERCADO ; 4 ; 2\n",
-    "EOLICA-SUBMERCADO ; 5 ; 2\n",
-    "",
-]
