+++ tests/mocks/arquivos/usinas_hidreletricas.py
@@ -1,0 +1,268 @@
+MockPolinjus = [
+    " &&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&\n",
+    " & Cadastro das familias de curvas de jusante\n",
+    " &\n",
+    " & Usina: codigo da usina hidraulica\n",
+    " & IndCurva: indice da familia (sequencial)\n",
+    " & HjusRef: nivel de montante da usina de jusante para referencia da familia\n",
+    " & nPol: quantidade de polinomios da familia\n",
+    " &\n",
+    " &&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&\n",
+    " &;Usina;Indice;HjusRef\n",
+    " &;Usina;Indice;nPol\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0001;001;  885.3052\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0001;001;002\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0001;002;  885.4000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0001;002;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0001;003;  885.7000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0001;003;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0001;004;  885.8000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0001;004;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0001;005;  886.0000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0001;005;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0002;001;  855.7895\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0002;001;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0004;001;  765.7641\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0004;001;002\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0004;002;  767.0000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0004;002;002\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0004;003;  767.3000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0004;003;002\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0004;004;  767.6000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0004;004;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0004;005;  768.0000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0004;005;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0006;001;  669.2448\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0006;001;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0007;001;  617.8585\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0007;001;002\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0007;002;  619.3000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0007;002;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0007;003;  620.9000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0007;003;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0007;004;  621.7000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0007;004;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0007;005;  622.5000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0007;005;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0008;001;  556.2537\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0008;001;002\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0008;002;  557.1000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0008;002;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0008;003;  557.4000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0008;003;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0008;004;  558.0000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0008;004;002\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0008;005;  558.5000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0008;005;002\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0009;001;  511.8859\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0009;001;002\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0009;002;  512.1000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0009;002;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0010;001;  493.5565\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0010;001;002\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0010;002;  494.1000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0010;002;003\n",
+    " &\n",
+    " HIDRELETRICA-CURVAJUSANTE                   ;0010;003;  494.6000\n",
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0010;003;003\n",
+    " &\n",
+    " &&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&\n",
+    " &\n",
+    " & Cadastro dos coeficientes dos polinômios por partes das curvas de jusante\n",
+    " &\n",
+    " & Usina: codigo da usina hidraulica\n",
+    " & IndCurva: indice da familia (sequencial)\n",
+    " & IndPolin Polinomio: indice do polinomio do segmento (sequencial)\n",
+    " & QjusMin: Limite inferior de vazao de jusante (defluencia mais lateral) para janela de validade do polinomio\n",
+    " & QjusMax: Limite superior de vazao de jusante (defluencia mais lateral) para janela de validade do polinomio\n",
+    " & a0 : coeficiente de grau 0 do polinomio\n",
+    " & a1 : coeficiente de grau 1 do polinomio\n",
+    " & a2 : coeficiente de grau 2 do polinomio\n",
+    " & a3 : coeficiente de grau 3 do polinomio\n",
+    " & a4 : coeficiente de grau 4 do polinomio\n",
+    " &\n",
+    " &&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&\n",
+    " &;Usina;IndCurva;IndPolin;QjusMin;QjusMax;a0;a1;a2;a3;a4\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;001; 1;               0.000;             408.649;0.88530520000000E+03;-.31521360000000E-17;0.19686530000000E-04;-.25518040000000E-07;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;001; 2;             408.649;            1054.722;0.88467330000000E+03;0.80666000000000E-02;-.86720890000000E-05;0.54010460000000E-08;-.13886850000000E-11\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;002; 1;               0.000;             123.055;0.88540000000000E+03;0.22264820000000E-02;0.72267780000000E-05;-.46752770000000E-06;0.28066640000000E-08\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;002; 2;             123.055;             408.649;0.88530520000000E+03;-.31521360000000E-17;0.19686530000000E-04;-.25518040000000E-07;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;002; 3;             408.649;            1054.722;0.88467330000000E+03;0.80666000000000E-02;-.86720890000000E-05;0.54010460000000E-08;-.13886850000000E-11\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;003; 1;               0.000;             396.401;0.88570000000000E+03;0.60492090000000E-03;0.26923990000000E-05;0.23324290000000E-07;-.40763360000000E-10\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;003; 2;             396.401;             408.649;0.88530520000000E+03;-.31521360000000E-17;0.19686530000000E-04;-.25518040000000E-07;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;003; 3;             408.649;            1054.722;0.88467330000000E+03;0.80666000000000E-02;-.86720890000000E-05;0.54010460000000E-08;-.13886850000000E-11\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;004; 1;               0.000;             365.407;0.88580000000000E+03;0.33844090000000E-18;0.88426950000000E-06;0.36830870000000E-07;-.57562770000000E-10\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;004; 2;             365.407;             408.649;0.88530520000000E+03;-.31521360000000E-17;0.19686530000000E-04;-.25518040000000E-07;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;004; 3;             408.649;            1054.722;0.88467330000000E+03;0.80666000000000E-02;-.86720890000000E-05;0.54010460000000E-08;-.13886850000000E-11\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;005; 1;               0.000;             365.400;0.88600000000000E+03;0.27478400000000E-03;-.64518770000000E-05;0.54412710000000E-07;-.67585460000000E-10\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;005; 2;             365.400;             408.649;0.88530520000000E+03;-.31521360000000E-17;0.19686530000000E-04;-.25518040000000E-07;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;005; 3;             408.649;            1054.722;0.88467330000000E+03;0.80666000000000E-02;-.86720890000000E-05;0.54010460000000E-08;-.13886850000000E-11\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0002;001; 1;               0.000;             242.693;0.85578947064344E+03;0.11562960551554E-01;0.49862043017189E-05;-.68537729932874E-07;0.21566855280797E-11\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0002;001; 2;             242.693;             863.797;0.85780957602716E+03;-.18541426936976E-02;0.11088364944033E-04;-.66823864597252E-08;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0002;001; 3;             863.797;            2363.670;0.85785060675746E+03;0.28525351307282E-02;0.87103266365734E-07;-.38914171007355E-09;0.82237970254910E-13\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;001; 1;               0.000;            1257.447;0.76576413085343E+03;0.11261094559232E-01;-.62425474410895E-05;0.17543098194994E-08;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;001; 2;            1257.447;            6238.234;0.76650542364729E+03;0.76767509717095E-02;-.19707200040794E-05;0.26917708438392E-09;-.14346039347404E-13\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;002; 1;               0.000;            1694.482;0.76700000000000E+03;0.74128152590876E-02;-.13055224042154E-05;-.64680229241739E-09;0.28880378718093E-12\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;002; 2;            1694.482;            6238.234;0.76650542364729E+03;0.76767509717095E-02;-.19707200040794E-05;0.26917708438392E-09;-.14346039347404E-13\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;003; 1;               0.000;            1351.670;0.76730000000000E+03;0.59631446780465E-02;0.22930377565992E-05;-.45129258347764E-08;0.16457046194850E-11\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;003; 2;            1351.670;            6238.234;0.76650542364729E+03;0.76767509717095E-02;-.19707200040794E-05;0.26917708438392E-09;-.14346039347404E-13\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;004; 1;               0.000;             842.406;0.76760000000000E+03;0.48674884293944E-02;0.60596706057144E-05;-.12708322384901E-07;0.68821071334384E-11\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;004; 2;             842.406;            1257.447;0.76576413085343E+03;0.11261094559232E-01;-.62425474410895E-05;0.17543098194994E-08;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;004; 3;            1257.447;            6238.234;0.76650542364729E+03;0.76767509717095E-02;-.19707200040794E-05;0.26917708438392E-09;-.14346039347404E-13\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;005; 1;               0.000;             869.181;0.76800000000000E+03;0.45961205658945E-02;-.45998700839511E-06;0.12952539683796E-08;-.89347859517125E-12\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;005; 2;             869.181;            1257.447;0.76576413085343E+03;0.11261094559232E-01;-.62425474410895E-05;0.17543098194994E-08;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0004;005; 3;            1257.447;            6238.234;0.76650542364729E+03;0.76767509717095E-02;-.19707200040794E-05;0.26917708438392E-09;-.14346039347404E-13\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0006;001; 1;               0.000;            1797.517;0.66924477454233E+03;0.61457290391761E-02;-.35372650006539E-05;0.76065725868450E-09;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0006;001; 2;            1797.517;            4190.058;0.65680447106812E+03;0.22214593230849E-01;-.10816507034675E-04;0.22954796490786E-08;-.17606898487237E-12\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0006;001; 3;            4190.058;           11245.494;0.66606835022913E+03;0.42007316246237E-02;-.74270446044706E-06;0.61461858612508E-10;-.18668560886631E-14\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;001; 1;               0.000;            4332.009;0.61774670000000E+03;0.25708320000000E-02;-.16324900000000E-06;0.00000000000000E+01;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;001; 2;            4332.009;           13081.880;0.62081030000000E+03;0.11564400000000E-02;-.18880750000000E-14;0.00000000000000E+01;0.00000000000000E+01\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;002; 1;               0.000;            1319.244;0.61946530000000E+03;0.17103310000000E-04;-.28092880000000E-06;0.15863900000000E-08;-.59000550000000E-12\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;002; 2;            1319.244;            4332.009;0.61774670000000E+03;0.25708320000000E-02;-.16324900000000E-06;0.00000000000000E+01;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;002; 3;            4332.009;           13081.880;0.62081030000000E+03;0.11564400000000E-02;-.18880750000000E-14;0.00000000000000E+01;0.00000000000000E+01\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;003; 1;               0.000;            3076.155;0.62071990000000E+03;-.13576460000000E-16;0.77548310000000E-06;-.20385050000000E-09;0.22178910000000E-13\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;003; 2;            3076.155;            4332.009;0.61774670000000E+03;0.25708320000000E-02;-.16324900000000E-06;0.00000000000000E+01;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;003; 3;            4332.009;           13081.880;0.62081030000000E+03;0.11564400000000E-02;-.18880750000000E-14;0.00000000000000E+01;0.00000000000000E+01\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;004; 1;               0.000;            3442.639;0.62156770000000E+03;-.13040120000000E-17;0.54572260000000E-06;-.13572050000000E-09;0.15409710000000E-13\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;004; 2;            3442.639;            4332.009;0.61774670000000E+03;0.25708320000000E-02;-.16324900000000E-06;0.00000000000000E+01;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;004; 3;            4332.009;           13081.880;0.62081030000000E+03;0.11564400000000E-02;-.18880750000000E-14;0.00000000000000E+01;0.00000000000000E+01\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;005; 1;               0.000;            3442.600;0.62220000000000E+03;-.22632450000000E-16;0.27779900000000E-06;-.42062500000000E-10;0.63090700000000E-14\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;005; 2;            3442.600;            4332.009;0.61774670000000E+03;0.25708320000000E-02;-.16324900000000E-06;0.00000000000000E+01;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0007;005; 3;            4332.009;           13081.880;0.62081030000000E+03;0.11564400000000E-02;-.18880750000000E-14;0.00000000000000E+01;0.00000000000000E+01\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;001; 1;               0.000;            4190.566;0.55625370000000E+03;0.23031610000000E-02;-.49837690000000E-06;0.48113230000000E-10;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;001; 2;            4190.566;           11989.420;0.55701400000000E+03;0.11675690000000E-02;-.88995910000000E-07;0.54029180000000E-11;-.15422390000000E-15\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;002; 1;               0.000;             869.428;0.55710000000000E+03;0.84437430000000E-17;0.73121700000000E-06;0.12095430000000E-08;-.93909760000000E-12\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;002; 2;             869.428;            4190.566;0.55625370000000E+03;0.23031610000000E-02;-.49837690000000E-06;0.48113230000000E-10;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;002; 3;            4190.566;           11989.420;0.55701400000000E+03;0.11675690000000E-02;-.88995910000000E-07;0.54029180000000E-11;-.15422390000000E-15\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;003; 1;               0.000;            1090.008;0.55740000000000E+03;0.24278880000000E-17;0.51480830000000E-07;0.13142440000000E-08;-.65797760000000E-12\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;003; 2;            1090.008;            4190.566;0.55625370000000E+03;0.23031610000000E-02;-.49837690000000E-06;0.48113230000000E-10;0.00000000000000E+01\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;003; 3;            4190.566;           11989.420;0.55701400000000E+03;0.11675690000000E-02;-.88995910000000E-07;0.54029180000000E-11;-.15422390000000E-15\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;004; 1;               0.000;            4378.874;0.55800000000000E+03;0.40780010000000E-18;0.67133030000000E-06;-.20616400000000E-09;0.19732290000000E-13\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;004; 2;            4378.874;           11989.420;0.55701400000000E+03;0.11675690000000E-02;-.88995910000000E-07;0.54029180000000E-11;-.15422390000000E-15\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;005; 1;               0.000;            4600.000;0.55850000000000E+03;-.84983960000000E-17;0.47108720000000E-06;-.13364260000000E-09;0.12280580000000E-13\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0008;005; 2;            4600.000;           11989.420;0.55701400000000E+03;0.11675690000000E-02;-.88995910000000E-07;0.54029180000000E-11;-.15422390000000E-15\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0009;001; 1;               0.000;            3573.979;0.51183311818750E+03;0.80007944081166E-03;0.55375691608354E-06;-.14126213644413E-09;0.94914919310503E-14\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0009;001; 2;            3573.979;           10957.615;0.51099815722606E+03;0.23714330034223E-02;-.25591616387685E-06;0.15935552601890E-10;-.40770261301443E-15\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0009;002; 1;               0.000;            1724.944;0.51204726008102E+03;-.99575241765259E-17;0.17931161426941E-05;-.93845372123431E-09;0.18681392886259E-12\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0009;002; 2;            1724.944;            3573.979;0.51183311818750E+03;0.80007944081166E-03;0.55375691608354E-06;-.14126213644413E-09;0.94914919310503E-14\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0009;002; 3;            3573.979;           10957.615;0.51099815722606E+03;0.23714330034223E-02;-.25591616387685E-06;0.15935552601890E-10;-.40770261301443E-15\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0010;001; 1;               0.000;            3931.811;4.93556500000000E+02;1.45685300000000E-03;3.82077800000000E-08;-7.6535160000000E-11;9.61812000000000E-15\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0010;001; 2;            3931.811;           11252.960;4.93953400000000E+02;1.39952600000000E-03;-1.6274530000000E-07;1.06834400000000E-11;-2.8370270000000E-16\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0010;002; 1;               0.000;             946.902;4.93572500000000E+02;2.09901800000000E-03;-1.3804510000000E-06;6.95680500000000E-10;0.00000000000000E+00\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0010;002; 2;             946.902;            3931.811;4.93556500000000E+02;1.45685300000000E-03;3.82077800000000E-08;-7.6535160000000E-11;9.61812000000000E-15\n",
+    "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0010;002; 3;            3931.811;           11252.960;4.93953400000000E+02;1.39952600000000E-03;-1.6274530000000E-07;1.06834400000000E-11;-2.8370270000000E-16\n",
+    " &\n",
+    "HIDRELETRICA-CURVAJUSANTE-AFOGAMENTO-EXPLICITO-USINA;0022;sim\n",
+    "HIDRELETRICA-CURVAJUSANTE-AFOGAMENTO-EXPLICITO-USINA;0044;sim\n",
+    "HIDRELETRICA-CURVAJUSANTE-AFOGAMENTO-EXPLICITO-USINA;0054;nao\n",
+    "HIDRELETRICA-CURVAJUSANTE-AFOGAMENTO-EXPLICITO-PADRAO;nao\n",
+]
+
+MockHidreletricaCurvaJusante = (
+    " HIDRELETRICA-CURVAJUSANTE                   ;0001;001;  885.3052"
+)
+
+MockHidreletricaCurvaJusantePolinomio = (
+    " HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES;0001;001;002"
+)
+
+MockHidreletricaCurvaJusantePolinomioSegmento = "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO;0001;001; 1;               0.000;             408.649;0.88530520000000E+03;-.31521360000000E-17;0.19686530000000E-04;-.25518040000000E-07;0.00000000000000E+01"
+
+MockHidreletricaCurvaJusanteAfogamentoExplicitoUsina = (
+    "HIDRELETRICA-CURVAJUSANTE-AFOGAMENTO-EXPLICITO-USINA;0022;sim"
+)
+
+MockHidreletricaCurvaJusanteAfogamentoExplicitoPadrao = (
+    "HIDRELETRICA-CURVAJUSANTE-AFOGAMENTO-EXPLICITO-PADRAO;nao"
+)
+
+
+MockEstacaoBombeamento = "ESTACAO-BOMBEAMENTO;      1; Sta Cecilia                  ;   125 ;   181 ;          0.20 ;    181"
+
+MockEstacaoBombeamentoLimitesPeriodoPatamar = "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      1 ;2023/03;       ;  1 ;             0 ;     181"
+
+MockEstacaoBombeamentoSubmercado = (
+    "ESTACAO-BOMBEAMENTO-SUBMERCADO;      1;       1"
+)
+
+MockEstacaoBombeamentoCompleto = [
+    "&                  ;CodBomb;NomeEstacaoBomb               ;UHEorig;UHEdest;ConsEstBomb    ;BombMax        \n",
+    "&******************;IIIIIII;SSSSSSSSSSSSSSSSSSSSSSSSSSSSSS;IIIIIII;IIIIIII;FFFFFFFFFFFFFFF;FFFFFFF\n",
+    "ESTACAO-BOMBEAMENTO;      1; Sta Cecilia                  ;   125 ;   181 ;          0.20 ;    181\n",
+    "ESTACAO-BOMBEAMENTO;      2; Vigario                      ;   181 ;   182 ;          0.44 ;    182\n",
+    "ESTACAO-BOMBEAMENTO;      3; Traicao                      ;   108 ;   109 ;          0.30 ;    109\n",
+    "ESTACAO-BOMBEAMENTO;      4; Pedreira                     ;   109 ;   118 ;          0.25 ;    118\n",
+    "&                                           ;CodBomb; DataI;  DataF ; Pat; BombMin       ; BombMax      \n",
+    "&******************************************;IIIIIIII;IIIIIII;IIIIIII;IIII;FFFFFFFFFFFFFFF;FFFFFFFF\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      1 ;2023/03;       ;  1 ;             0 ;     181\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      1 ;2023/03;       ;  2 ;             0 ;     181\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      1 ;2023/03;       ;  3 ;             0 ;     181\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      2 ;2023/03;       ;  1 ;             0 ;     182\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      2 ;2023/03;       ;  2 ;             0 ;     182\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      2 ;2023/03;       ;  3 ;             0 ;     182\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      3 ;2023/03;       ;  1 ;             0 ;     109\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      3 ;2023/03;       ;  2 ;             0 ;     109\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      3 ;2023/03;       ;  3 ;             0 ;     109\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      4 ;2023/03;       ;  1 ;             0 ;     118\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      4 ;2023/03;       ;  2 ;             0 ;     118\n",
+    "ESTACAO-BOMBEAMENTO-LIMITES-PERIODO-PATAMAR;      4 ;2023/03;       ;  3 ;             0 ;     118\n",
+    "&                             ;CodBomb;CodSubm ;\n",
+    "&*****************************;IIIIIII;IIIIIIII;\n",
+    "ESTACAO-BOMBEAMENTO-SUBMERCADO;      1;       1\n",
+    "ESTACAO-BOMBEAMENTO-SUBMERCADO;      2;       1\n",
+    "ESTACAO-BOMBEAMENTO-SUBMERCADO;      3;       1\n",
+    "ESTACAO-BOMBEAMENTO-SUBMERCADO;      4;       1\n",
+]
+
+MockUsinasHidreletricas = MockPolinjus + MockEstacaoBombeamentoCompleto
